+++ src/Symfony/Bridge/Doctrine/Tests/Validator/Constraints/UniqueEntityValidatorTest.php
@@ -84,10 +84,16 @@ class UniqueEntityValidatorTest extends ConstraintValidatorTestCase
     protected function createRegistryMock($em = null)
     {
         $registry = $this->createMock(ManagerRegistry::class);
-        $registry->expects($this->any())
-                 ->method('getManager')
-                 ->with($this->equalTo(self::EM_NAME))
-                 ->willReturn($em);
+
+        if (null === $em) {
+            $registry->method('getManager')
+                ->with($this->equalTo(self::EM_NAME))
+                ->willThrowException(new \InvalidArgumentException());
+        } else {
+            $registry->method('getManager')
+                ->with($this->equalTo(self::EM_NAME))
+                ->willReturn($em);
+        }
 
         return $registry;
     }
@@ -165,11 +171,11 @@ class UniqueEntityValidatorTest extends ConstraintValidatorTestCase
 
     /**
      * This is a functional test as there is a large integration necessary to get the validator working.
-     *
-     * @dataProvider provideUniquenessConstraints
      */
-    public function testValidateUniqueness(UniqueEntity $constraint)
+    public function testValidateUniqueness()
     {
+        $constraint = new UniqueEntity(message: 'myMessage', fields: ['name'], em: 'foo');
+
         $entity1 = new SingleIntIdEntity(1, 'Foo');
         $entity2 = new SingleIntIdEntity(2, 'Foo');
 
@@ -217,6 +223,28 @@ class UniqueEntityValidatorTest extends ConstraintValidatorTestCase
         // this will load a proxy object
         $entity = $this->em->getReference(SingleIntIdWithPrivateNameEntity::class, 1);
 
+        $this->validator->validate($entity, new UniqueEntity(
+            fields: ['name'],
+            em: self::EM_NAME,
+        ));
+
+        $this->assertNoViolation();
+    }
+
+    /**
+     * @group legacy
+     */
+    public function testValidateEntityWithPrivatePropertyAndProxyObjectDoctrineStyle()
+    {
+        $entity = new SingleIntIdWithPrivateNameEntity(1, 'Foo');
+        $this->em->persist($entity);
+        $this->em->flush();
+
+        $this->em->clear();
+
+        // this will load a proxy object
+        $entity = $this->em->getReference(SingleIntIdWithPrivateNameEntity::class, 1);
+
         $this->validator->validate($entity, new UniqueEntity([
             'fields' => ['name'],
             'em' => self::EM_NAME,
@@ -225,10 +253,7 @@ class UniqueEntityValidatorTest extends ConstraintValidatorTestCase
         $this->assertNoViolation();
     }
 
-    /**
-     * @dataProvider provideConstraintsWithCustomErrorPath
-     */
-    public function testValidateCustomErrorPath(UniqueEntity $constraint)
+    public function testValidateCustomErrorPath()
     {
         $entity1 = new SingleIntIdEntity(1, 'Foo');
         $entity2 = new SingleIntIdEntity(2, 'Foo');
@@ -236,7 +261,7 @@ class UniqueEntityValidatorTest extends ConstraintValidatorTestCase
         $this->em->persist($entity1);
         $this->em->flush();
 
-        $this->validator->validate($entity2, $constraint);
+        $this->validator->validate($entity2, new UniqueEntity(message: 'myMessage', fields: ['name'], em: 'foo', errorPath: 'bar'));
 
         $this->buildViolation('myMessage')
             ->atPath('property.path.bar')
@@ -247,22 +272,34 @@ class UniqueEntityValidatorTest extends ConstraintValidatorTestCase
             ->assertRaised();
     }
 
-    public static function provideConstraintsWithCustomErrorPath(): iterable
+    /**
+     * @group legacy
+     */
+    public function testValidateCustomErrorPathDoctrineStyle()
     {
-        yield 'Doctrine style' => [new UniqueEntity([
+        $entity1 = new SingleIntIdEntity(1, 'Foo');
+        $entity2 = new SingleIntIdEntity(2, 'Foo');
+
+        $this->em->persist($entity1);
+        $this->em->flush();
+
+        $this->validator->validate($entity2, new UniqueEntity([
             'message' => 'myMessage',
             'fields' => ['name'],
-            'em' => self::EM_NAME,
+            'em' => 'foo',
             'errorPath' => 'bar',
-        ])];
+        ]));
 
-        yield 'Named arguments' => [new UniqueEntity(message: 'myMessage', fields: ['name'], em: 'foo', errorPath: 'bar')];
+        $this->buildViolation('myMessage')
+            ->atPath('property.path.bar')
+            ->setParameter('{{ value }}', '"Foo"')
+            ->setInvalidValue($entity2)
+            ->setCause([$entity1])
+            ->setCode(UniqueEntity::NOT_UNIQUE_ERROR)
+            ->assertRaised();
     }
 
-    /**
-     * @dataProvider provideUniquenessConstraints
-     */
-    public function testValidateUniquenessWithNull(UniqueEntity $constraint)
+    public function testValidateUniquenessWithNull()
     {
         $entity1 = new SingleIntIdEntity(1, null);
         $entity2 = new SingleIntIdEntity(2, null);
@@ -271,7 +308,7 @@ class UniqueEntityValidatorTest extends ConstraintValidatorTestCase
         $this->em->persist($entity2);
         $this->em->flush();
 
-        $this->validator->validate($entity1, $constraint);
+        $this->validator->validate($entity1, new UniqueEntity(message: 'myMessage', fields: ['name'], em: 'foo'));
 
         $this->assertNoViolation();
     }
@@ -309,13 +346,6 @@ class UniqueEntityValidatorTest extends ConstraintValidatorTestCase
 
     public static function provideConstraintsWithIgnoreNullDisabled(): iterable
     {
-        yield 'Doctrine style' => [new UniqueEntity([
-            'message' => 'myMessage',
-            'fields' => ['name', 'name2'],
-            'em' => self::EM_NAME,
-            'ignoreNull' => false,
-        ])];
-
         yield 'Named arguments' => [new UniqueEntity(message: 'myMessage', fields: ['name', 'name2'], em: 'foo', ignoreNull: false)];
     }
 
@@ -357,36 +387,22 @@ class UniqueEntityValidatorTest extends ConstraintValidatorTestCase
 
     public static function provideConstraintsWithIgnoreNullEnabled(): iterable
     {
-        yield 'Doctrine style' => [new UniqueEntity([
-            'message' => 'myMessage',
-            'fields' => ['name', 'name2'],
-            'em' => self::EM_NAME,
-            'ignoreNull' => true,
-        ])];
-
         yield 'Named arguments' => [new UniqueEntity(message: 'myMessage', fields: ['name', 'name2'], em: 'foo', ignoreNull: true)];
     }
 
     public static function provideConstraintsWithIgnoreNullEnabledOnFirstField(): iterable
     {
-        yield 'Doctrine style (name field)' => [new UniqueEntity([
-            'message' => 'myMessage',
-            'fields' => ['name', 'name2'],
-            'em' => self::EM_NAME,
-            'ignoreNull' => 'name',
-        ])];
-
         yield 'Named arguments (name field)' => [new UniqueEntity(message: 'myMessage', fields: ['name', 'name2'], em: 'foo', ignoreNull: 'name')];
     }
 
     public function testValidateUniquenessWithValidCustomErrorPath()
     {
-        $constraint = new UniqueEntity([
-            'message' => 'myMessage',
-            'fields' => ['name', 'name2'],
-            'em' => self::EM_NAME,
-            'errorPath' => 'name2',
-        ]);
+        $constraint = new UniqueEntity(
+            message: 'myMessage',
+            fields: ['name', 'name2'],
+            em: self::EM_NAME,
+            errorPath: 'name2',
+        );
 
         $entity1 = new DoubleNameEntity(1, 'Foo', 'Bar');
         $entity2 = new DoubleNameEntity(2, 'Foo', 'Bar');
@@ -413,10 +429,7 @@ class UniqueEntityValidatorTest extends ConstraintValidatorTestCase
             ->assertRaised();
     }
 
-    /**
-     * @dataProvider provideConstraintsWithCustomRepositoryMethod
-     */
-    public function testValidateUniquenessUsingCustomRepositoryMethod(UniqueEntity $constraint)
+    public function testValidateUniquenessUsingCustomRepositoryMethod()
     {
         $repository = $this->createRepositoryMock(SingleIntIdEntity::class);
         $repository->expects($this->once())
@@ -430,15 +443,12 @@ class UniqueEntityValidatorTest extends ConstraintValidatorTestCase
 
         $entity1 = new SingleIntIdEntity(1, 'foo');
 
-        $this->validator->validate($entity1, $constraint);
+        $this->validator->validate($entity1, new UniqueEntity(message: 'myMessage', fields: ['name'], em: 'foo', repositoryMethod: 'findByCustom'));
 
         $this->assertNoViolation();
     }
 
-    /**
-     * @dataProvider provideConstraintsWithCustomRepositoryMethod
-     */
-    public function testValidateUniquenessWithUnrewoundArray(UniqueEntity $constraint)
+    public function testValidateUniquenessWithUnrewoundArray()
     {
         $entity = new SingleIntIdEntity(1, 'foo');
 
@@ -461,21 +471,9 @@ class UniqueEntityValidatorTest extends ConstraintValidatorTestCase
         $this->validator = $this->createValidator();
         $this->validator->initialize($this->context);
 
-        $this->validator->validate($entity, $constraint);
+        $this->validator->validate($entity, new UniqueEntity(message: 'myMessage', fields: ['name'], em: 'foo', repositoryMethod: 'findByCustom'));
 
         $this->assertNoViolation();
-    }
-
-    public static function provideConstraintsWithCustomRepositoryMethod(): iterable
-    {
-        yield 'Doctrine style' => [new UniqueEntity([
-            'message' => 'myMessage',
-            'fields' => ['name'],
-            'em' => self::EM_NAME,
-            'repositoryMethod' => 'findByCustom',
-        ])];
-
-        yield 'Named arguments' => [new UniqueEntity(message: 'myMessage', fields: ['name'], em: 'foo', repositoryMethod: 'findByCustom')];
     }
 
     /**
@@ -483,12 +481,12 @@ class UniqueEntityValidatorTest extends ConstraintValidatorTestCase
      */
     public function testValidateResultTypes($entity1, $result)
     {
-        $constraint = new UniqueEntity([
-            'message' => 'myMessage',
-            'fields' => ['name'],
-            'em' => self::EM_NAME,
-            'repositoryMethod' => 'findByCustom',
-        ]);
+        $constraint = new UniqueEntity(
+            message: 'myMessage',
+            fields: ['name'],
+            em: self::EM_NAME,
+            repositoryMethod: 'findByCustom',
+        );
 
         $repository = $this->createRepositoryMock($entity1::class);
         $repository->expects($this->once())
@@ -518,11 +516,11 @@ class UniqueEntityValidatorTest extends ConstraintValidatorTestCase
 
     public function testAssociatedEntity()
     {
-        $constraint = new UniqueEntity([
-            'message' => 'myMessage',
-            'fields' => ['single'],
-            'em' => self::EM_NAME,
-        ]);
+        $constraint = new UniqueEntity(
+            message: 'myMessage',
+            fields: ['single'],
+            em: self::EM_NAME,
+        );
 
         $entity1 = new SingleIntIdEntity(1, 'foo');
         $associated = new AssociationEntity();
@@ -554,11 +552,11 @@ class UniqueEntityValidatorTest extends ConstraintValidatorTestCase
 
     public function testValidateUniquenessNotToStringEntityWithAssociatedEntity()
     {
-        $constraint = new UniqueEntity([
-            'message' => 'myMessage',
-            'fields' => ['single'],
-            'em' => self::EM_NAME,
-        ]);
+        $constraint = new UniqueEntity(
+            message: 'myMessage',
+            fields: ['single'],
+            em: self::EM_NAME,
+        );
 
         $entity1 = new SingleIntIdNoToStringEntity(1, 'foo');
         $associated = new AssociationEntity2();
@@ -592,12 +590,12 @@ class UniqueEntityValidatorTest extends ConstraintValidatorTestCase
 
     public function testAssociatedEntityWithNull()
     {
-        $constraint = new UniqueEntity([
-            'message' => 'myMessage',
-            'fields' => ['single'],
-            'em' => self::EM_NAME,
-            'ignoreNull' => false,
-        ]);
+        $constraint = new UniqueEntity(
+            message: 'myMessage',
+            fields: ['single'],
+            em: self::EM_NAME,
+            ignoreNull: false,
+        );
 
         $associated = new AssociationEntity();
         $associated->single = null;
@@ -649,12 +647,12 @@ class UniqueEntityValidatorTest extends ConstraintValidatorTestCase
         $repository = $this->createRepositoryMock(SingleIntIdEntity::class);
         $this->repositoryFactory->setRepository($this->em, SingleIntIdEntity::class, $repository);
 
-        $constraint = new UniqueEntity([
-            'message' => 'myMessage',
-            'fields' => ['phoneNumbers'],
-            'em' => self::EM_NAME,
-            'repositoryMethod' => 'findByCustom',
-        ]);
+        $constraint = new UniqueEntity(
+            message: 'myMessage',
+            fields: ['phoneNumbers'],
+            em: self::EM_NAME,
+            repositoryMethod: 'findByCustom',
+        );
 
         $entity1 = new SingleIntIdEntity(1, 'foo');
         $entity1->phoneNumbers[] = 123;
@@ -685,11 +683,11 @@ class UniqueEntityValidatorTest extends ConstraintValidatorTestCase
 
     public function testDedicatedEntityManagerNullObject()
     {
-        $constraint = new UniqueEntity([
-            'message' => 'myMessage',
-            'fields' => ['name'],
-            'em' => self::EM_NAME,
-        ]);
+        $constraint = new UniqueEntity(
+            message: 'myMessage',
+            fields: ['name'],
+            em: self::EM_NAME,
+        );
 
         $this->em = null;
         $this->registry = $this->createRegistryMock($this->em);
@@ -706,11 +704,11 @@ class UniqueEntityValidatorTest extends ConstraintValidatorTestCase
 
     public function testEntityManagerNullObject()
     {
-        $constraint = new UniqueEntity([
-            'message' => 'myMessage',
-            'fields' => ['name'],
+        $constraint = new UniqueEntity(
+            message: 'myMessage',
+            fields: ['name'],
             // no "em" option set
-        ]);
+        );
 
         $this->em = null;
         $this->registry = $this->createRegistryMock($this->em);
@@ -738,11 +736,11 @@ class UniqueEntityValidatorTest extends ConstraintValidatorTestCase
         $this->validator = $this->createValidator();
         $this->validator->initialize($this->context);
 
-        $constraint = new UniqueEntity([
-            'message' => 'myMessage',
-            'fields' => ['name'],
-            'em' => self::EM_NAME,
-        ]);
+        $constraint = new UniqueEntity(
+            message: 'myMessage',
+            fields: ['name'],
+            em: self::EM_NAME,
+        );
 
         $entity = new SingleIntIdEntity(1, null);
 
@@ -755,12 +753,12 @@ class UniqueEntityValidatorTest extends ConstraintValidatorTestCase
 
     public function testValidateInheritanceUniqueness()
     {
-        $constraint = new UniqueEntity([
-            'message' => 'myMessage',
-            'fields' => ['name'],
-            'em' => self::EM_NAME,
-            'entityClass' => Person::class,
-        ]);
+        $constraint = new UniqueEntity(
+            message: 'myMessage',
+            fields: ['name'],
+            em: self::EM_NAME,
+            entityClass: Person::class,
+        );
 
         $entity1 = new Person(1, 'Foo');
         $entity2 = new Employee(2, 'Foo');
@@ -789,12 +787,12 @@ class UniqueEntityValidatorTest extends ConstraintValidatorTestCase
 
     public function testInvalidateRepositoryForInheritance()
     {
-        $constraint = new UniqueEntity([
-            'message' => 'myMessage',
-            'fields' => ['name'],
-            'em' => self::EM_NAME,
-            'entityClass' => SingleStringIdEntity::class,
-        ]);
+        $constraint = new UniqueEntity(
+            message: 'myMessage',
+            fields: ['name'],
+            em: self::EM_NAME,
+            entityClass: SingleStringIdEntity::class,
+        );
 
         $entity = new Person(1, 'Foo');
 
@@ -806,11 +804,11 @@ class UniqueEntityValidatorTest extends ConstraintValidatorTestCase
 
     public function testValidateUniquenessWithCompositeObjectNoToStringIdEntity()
     {
-        $constraint = new UniqueEntity([
-            'message' => 'myMessage',
-            'fields' => ['objectOne', 'objectTwo'],
-            'em' => self::EM_NAME,
-        ]);
+        $constraint = new UniqueEntity(
+            message: 'myMessage',
+            fields: ['objectOne', 'objectTwo'],
+            em: self::EM_NAME,
+        );
 
         $objectOne = new SingleIntIdNoToStringEntity(1, 'foo');
         $objectTwo = new SingleIntIdNoToStringEntity(2, 'bar');
@@ -841,11 +839,11 @@ class UniqueEntityValidatorTest extends ConstraintValidatorTestCase
 
     public function testValidateUniquenessWithCustomDoctrineTypeValue()
     {
-        $constraint = new UniqueEntity([
-            'message' => 'myMessage',
-            'fields' => ['name'],
-            'em' => self::EM_NAME,
-        ]);
+        $constraint = new UniqueEntity(
+            message: 'myMessage',
+            fields: ['name'],
+            em: self::EM_NAME,
+        );
 
         $existingEntity = new SingleIntIdStringWrapperNameEntity(1, new StringWrapper('foo'));
 
@@ -872,11 +870,11 @@ class UniqueEntityValidatorTest extends ConstraintValidatorTestCase
      */
     public function testValidateUniquenessCause()
     {
-        $constraint = new UniqueEntity([
-            'message' => 'myMessage',
-            'fields' => ['name'],
-            'em' => self::EM_NAME,
-        ]);
+        $constraint = new UniqueEntity(
+            message: 'myMessage',
+            fields: ['name'],
+            em: self::EM_NAME,
+        );
 
         $entity1 = new SingleIntIdEntity(1, 'Foo');
         $entity2 = new SingleIntIdEntity(2, 'Foo');
@@ -908,12 +906,12 @@ class UniqueEntityValidatorTest extends ConstraintValidatorTestCase
      */
     public function testValidateUniquenessWithEmptyIterator($entity, $result)
     {
-        $constraint = new UniqueEntity([
-            'message' => 'myMessage',
-            'fields' => ['name'],
-            'em' => self::EM_NAME,
-            'repositoryMethod' => 'findByCustom',
-        ]);
+        $constraint = new UniqueEntity(
+            message: 'myMessage',
+            fields: ['name'],
+            em: self::EM_NAME,
+            repositoryMethod: 'findByCustom',
+        );
 
         $repository = $this->createRepositoryMock($entity::class);
         $repository->expects($this->once())
@@ -932,11 +930,11 @@ class UniqueEntityValidatorTest extends ConstraintValidatorTestCase
 
     public function testValueMustBeObject()
     {
-        $constraint = new UniqueEntity([
-            'message' => 'myMessage',
-            'fields' => ['name'],
-            'em' => self::EM_NAME,
-        ]);
+        $constraint = new UniqueEntity(
+            message: 'myMessage',
+            fields: ['name'],
+            em: self::EM_NAME,
+        );
 
         $this->expectException(UnexpectedValueException::class);
 
@@ -945,11 +943,11 @@ class UniqueEntityValidatorTest extends ConstraintValidatorTestCase
 
     public function testValueCanBeNull()
     {
-        $constraint = new UniqueEntity([
-            'message' => 'myMessage',
-            'fields' => ['name'],
-            'em' => self::EM_NAME,
-        ]);
+        $constraint = new UniqueEntity(
+            message: 'myMessage',
+            fields: ['name'],
+            em: self::EM_NAME,
+        );
 
         $this->validator->validate(null, $constraint);
 
@@ -1046,6 +1044,9 @@ class UniqueEntityValidatorTest extends ConstraintValidatorTestCase
             ->assertRaised();
     }
 
+    /**
+     * @group legacy
+     */
     public function testValidateDTOUniquenessDoctrineStyle()
     {
         $constraint = new UniqueEntity([
@@ -1106,6 +1107,9 @@ class UniqueEntityValidatorTest extends ConstraintValidatorTestCase
             ->assertRaised();
     }
 
+    /**
+     * @group legacy
+     */
     public function testValidateMappingOfFieldNamesDoctrineStyle()
     {
         $constraint = new UniqueEntity([
@@ -1147,6 +1151,9 @@ class UniqueEntityValidatorTest extends ConstraintValidatorTestCase
         $this->validator->validate($dto, $constraint);
     }
 
+    /**
+     * @group legacy
+     */
     public function testInvalidateDTOFieldNameDoctrineStyle()
     {
         $this->expectException(ConstraintDefinitionException::class);
@@ -1177,6 +1184,9 @@ class UniqueEntityValidatorTest extends ConstraintValidatorTestCase
         $this->validator->validate($dto, $constraint);
     }
 
+    /**
+     * @group legacy
+     */
     public function testInvalidateEntityFieldNameDoctrineStyle()
     {
         $this->expectException(ConstraintDefinitionException::class);
@@ -1222,6 +1232,9 @@ class UniqueEntityValidatorTest extends ConstraintValidatorTestCase
             ->assertRaised();
     }
 
+    /**
+     * @group legacy
+     */
     public function testValidateDTOUniquenessWhenUpdatingEntityDoctrineStyle()
     {
         $constraint = new UniqueEntity([
@@ -1274,6 +1287,9 @@ class UniqueEntityValidatorTest extends ConstraintValidatorTestCase
         $this->assertNoViolation();
     }
 
+    /**
+     * @group legacy
+     */
     public function testValidateDTOUniquenessWhenUpdatingEntityWithTheSameValueDoctrineStyle()
     {
         $constraint = new UniqueEntity([
@@ -1325,6 +1341,9 @@ class UniqueEntityValidatorTest extends ConstraintValidatorTestCase
         $this->assertNoViolation();
     }
 
+    /**
+     * @group legacy
+     */
     public function testValidateIdentifierMappingOfFieldNamesDoctrineStyle()
     {
         $constraint = new UniqueEntity([
@@ -1382,6 +1401,9 @@ class UniqueEntityValidatorTest extends ConstraintValidatorTestCase
         $this->validator->validate($dto, $constraint);
     }
 
+    /**
+     * @group legacy
+     */
     public function testInvalidateMissingIdentifierFieldNameDoctrineStyle()
     {
         $this->expectException(ConstraintDefinitionException::class);
@@ -1429,6 +1451,9 @@ class UniqueEntityValidatorTest extends ConstraintValidatorTestCase
         $this->validator->validate($dto, $constraint);
     }
 
+    /**
+     * @group legacy
+     */
     public function testUninitializedValueThrowExceptionDoctrineStyle()
     {
         $this->expectExceptionMessage('Typed property Symfony\Bridge\Doctrine\Tests\Fixtures\Dto::$foo must not be accessed before initialization');
@@ -1469,6 +1494,9 @@ class UniqueEntityValidatorTest extends ConstraintValidatorTestCase
         $this->validator->validate($dto, $constraint);
     }
 
+    /**
+     * @group legacy
+     */
     public function testEntityManagerNullObjectWhenDTODoctrineStyle()
     {
         $this->expectException(ConstraintDefinitionException::class);