+++ src/Symfony/Bundle/SecurityBundle/Tests/Functional/AccessTokenTest.php
@@ -425,7 +425,7 @@ class AccessTokenTest extends AbstractWebTestCase
         $this->assertSame(['message' => 'Welcome @dunglas!'], json_decode($response->getContent(), true));
     }
 
-    public function validAccessTokens(): array
+    public static function validAccessTokens(): array
     {
         if (!\extension_loaded('openssl')) {
             return [];
@@ -440,8 +440,8 @@ class AccessTokenTest extends AbstractWebTestCase
             'sub' => 'e21bf182-1538-406e-8ccb-e25a17aba39f',
             'username' => 'dunglas',
         ];
-        $jws = $this->createJws($claims);
-        $jwe = $this->createJwe($jws);
+        $jws = self::createJws($claims);
+        $jwe = self::createJwe($jws);
 
         return [
             [$jws],
@@ -509,7 +509,7 @@ class AccessTokenTest extends AbstractWebTestCase
         return (new JweCompactSerializer())->serialize(
             (new JWEBuilder(new AlgorithmManager([
                 new ECDHES(), new A128GCM(),
-            ])))->create()
+            ]), null))->create()
                 ->withPayload($input)
                 ->withSharedProtectedHeader(['alg' => 'ECDH-ES', 'enc' => 'A128GCM', ...$header])
                 // tip: use https://mkjwk.org/ to generate a JWK